% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agregar_titulo_nota.R
\name{agregar_titulo_nota}
\alias{agregar_titulo_nota}
\title{Agregar título y nota a un data frame o tibble (deprecated)}
\usage{
agregar_titulo_nota(tabla, titulo = "", nota = "")
}
\arguments{
\item{tabla}{Data frame o tibble sobre el que se añadirá el título y/o la nota.}

\item{titulo}{Cadena de texto que representa el título. Por defecto, cadena vacía.}

\item{nota}{Cadena de texto que representa la nota. Por defecto, cadena vacía.}
}
\value{
Data frame o tibble modificado que contiene la columna y/o fila con título y nota.
}
\description{
Esta función agrega una columna denominada "Titulo" al inicio del objeto \code{tabla}
si se especifica un título, y añade una fila con la nota al final si se proporciona una.
}
\details{
La función desagrupa el objeto \code{tabla} para evitar problemas de agrupación. Si se
provee un título, se crea una nueva columna llamada "Titulo" que se inserta al inicio del data frame;
si se provee una nota, se agrega como una fila adicional al final. Se asignan atributos al objeto
resultante para conservar estos valores.
}
\note{
\strong{Deprecated}  Esta función está obsoleta y será eliminada en futuras actualizaciones. Utiliza las funciones \verb{crear_Excel_*()} en su lugar.
}
\examples{
\dontrun{
  library(dplyr)
  library(tibble)
  t <- tibble(x = 1:5, y = 6:10)
  t_mod <- agregar_titulo_nota(tabla = t, titulo = "Datos", nota = "Fuente: encuesta")
}

}
