% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisisDescriptivo-package.R
\docType{package}
\name{analisisDescriptivo-package}
\alias{analisisDescriptivo}
\alias{analisisDescriptivo-package}
\title{analisisDescriptivo: What the Package Does (One Line, Title Case)}
\description{
Funciones minimas para realizar una analisis descriptivo basico.
}
\author{
\strong{Maintainer}: Sonia Plasencia-Carillo \email{splasenci3@alumno.uned.es}

}
\keyword{internal}
