% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asignar_filas_tablas.R
\name{asignar_filas_tablas}
\alias{asignar_filas_tablas}
\title{Asigna filas para la exportación de tablas a Excel}
\usage{
asignar_filas_tablas(list_tablas)
}
\arguments{
\item{list_tablas}{Lista de data.frames que se exportarán.}
}
\value{
Un tibble con las siguientes columnas:
\describe{
\item{Tablas}{Nombre asignado a cada tabla.}
\item{Nfilas}{Número de filas de cada tabla.}
\item{primeraFila}{Fila de inicio para cada tabla en el Excel.}
\item{tituloFila}{Fila en la que se colocará el título de la tabla (generalmente una fila antes de la tabla).}
}
}
\description{
Esta función determina la posición (filas) en la que se colocarán cada tabla y su título en un archivo Excel.
La primera tabla se escribe a partir de la fila 4 y las siguientes se separan por 3 filas adicionales.
}
\examples{
list_temp <- list(head(mtcars,5),head(mtcars,5),head(mtcars,5))
asignar_filas_tablas( list_temp )

}
