% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{aux_suma}
\alias{aux_suma}
\title{Funcion auxiliar}
\usage{
aux_suma(x)
}
\arguments{
\item{x}{lista de los valores dentro de la columna del data frame}
}
\value{
data frame con el \code{n_validos} y \code{n_NAs} que hay en una columna
}
\description{
Ayuda a la funcion suma_valores() a calcular el numero de valores validos y NAs
}
