% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{cal_descript}
\alias{cal_descript}
\title{Calculo descriptivos del data frame}
\usage{
cal_descript(tabla, columnas = NULL)
}
\arguments{
\item{tabla}{Data frame con varias columnas numericas para cada muestra}

\item{columnas}{Valor opcional \code{columnas} para filtrar por la(s) variable(s)
que se quieren analizar}
}
\value{
Data frame con los descriptivos (Media, Mediana, Desviacion, Minimo y Maximo)
}
\description{
Calcula los estadisticos descriptivos mas comunes de todas las columnas numericas
del data frame
}
