% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{calc_freq}
\alias{calc_freq}
\title{Calculo de frecuencias}
\usage{
calc_freq(tabla, columnas)
}
\arguments{
\item{tabla}{Data frame con varias columnas numericas para cada muestra}

\item{columnas}{Valor opcional \code{columnas} para filtrar por la(s) variable(s)
que se quieren analizar}
}
\value{
Data frame con la frecuencia que aparece cada elemento en cada columna
}
\description{
Calcula la frecuencia cn la que un elemento aparece en cada columna del Data frame
}
