% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coeficiente_correlacion_ponderado.R
\name{coeficiente_correlacion_ponderado}
\alias{coeficiente_correlacion_ponderado}
\title{Calcula el coeficiente de correlación ponderado}
\usage{
coeficiente_correlacion_ponderado(x, y, pesos)
}
\arguments{
\item{x}{Un vector numérico.}

\item{y}{Un vector numérico.}

\item{pesos}{Un vector numérico de pesos.}
}
\value{
Un valor numérico que representa el coeficiente de correlación ponderado.
}
\description{
Esta función calcula el coeficiente de correlación ponderado entre dos vectores numéricos,
utilizando los pesos correspondientes.
}
\examples{
\dontrun{
temp <- mtcars
temp$w <- rlnorm(nrow(temp))
coeficiente_correlacion_ponderado( temp$mpg, temp$disp, pesos = temp$w  )
cor( temp$mpg, temp$disp  )
}

}
