% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertir_porcentaje_a_numero.R
\name{convert_p_num}
\alias{convert_p_num}
\title{Convertir porcentaje en formato texto a numérico}
\usage{
convert_p_num(x)
}
\arguments{
\item{x}{Vector de caracteres representando porcentajes (ejemplo: "12,5\%").}
}
\value{
Vector numérico.
}
\description{
Elimina el símbolo de porcentaje y reemplaza la coma decimal por un punto, para finalmente convertir la cadena en numérico.
}
\examples{
convert_p_num(c("12,5\%", "100,0\%"))

}
