% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_p.R
\name{count_p}
\alias{count_p}
\title{Calcular Conteo y Proporciones}
\usage{
count_p(df, ..., wt = NULL)
}
\arguments{
\item{df}{Data frame. Conjunto de datos en el que se realizará el conteo.}

\item{...}{Variables de agrupación o expresiones sobre las cuales se calculará el conteo.}
}
\value{
Un tibble con las columnas resultantes de \code{dplyr::count()} (incluyendo \code{n} para el conteo)
y una columna \code{p} que contiene la proporción de cada grupo calculada como \code{n/sum(n)}.
}
\description{
Esta función realiza un conteo de ocurrencias en un data frame según las variables especificadas y
añade una columna con la proporción de cada grupo respecto al total.
}
\details{
La función utiliza \code{dplyr::count()} para obtener los conteos y posteriormente \code{dplyr::mutate()}
para calcular la proporción de cada grupo respecto al total. Es una manera práctica de combinar ambos pasos en uno solo.
}
\examples{
library(dplyr)
# Ejemplo con un data frame simple:
df <- tibble(grupo = c("A", "A", "B", "B", "B"))
count_p(df, grupo)

}
