% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covarianza_ponderada.R
\name{covarianza_ponderada}
\alias{covarianza_ponderada}
\title{Calcula la covarianza ponderada}
\usage{
covarianza_ponderada(x, y, pesos)
}
\arguments{
\item{x}{Un vector numérico.}

\item{y}{Un vector numérico.}

\item{pesos}{Un vector numérico de pesos.}
}
\value{
Un valor numérico que representa la covarianza ponderada.
}
\description{
Esta función calcula la covarianza ponderada entre dos vectores numéricos, considerando un vector
de pesos. Los valores de los vectores de entrada que son NA se eliminan automáticamente.
}
\examples{
\dontrun{
temp <- mtcars
temp$w <- rlnorm(nrow(temp))
covarianza_ponderada( temp$mpg, temp$disp, pesos = temp$w  )
cov( temp$mpg, temp$disp  )
}

}
