% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_Excel_tabla.R
\name{crear_Excel_tabla}
\alias{crear_Excel_tabla}
\title{Crea una hoja de Excel con una tabla formateada}
\usage{
crear_Excel_tabla(
  tabla,
  workbook = NULL,
  nombre_hoja = NULL,
  hoja = 1,
  formatear_tabla = TRUE,
  tipo_tabla = NULL,
  hay_var_grupo = NULL,
  titulo_tabla = NULL,
  numero_fila = NULL,
  titulo_principal = NULL,
  fuente = "Calibri",
  titulo_tabla_colorFuente = "#00b2a9",
  titulo_tabla_tamanyo = 12,
  estilo_tabla = "TableStyleLight1",
  tabla_conFiltro = T,
  titulo_principal_tamanyo = 14,
  titulo_principal_colorFuente = "#00b2a9",
  nombre_archivo = NULL,
  exportar = FALSE,
  sobreescribir_archivo = TRUE
)
}
\arguments{
\item{tabla}{Data.frame o tibble. La tabla de datos a exportar.}

\item{workbook}{Objeto workbook de \code{openxlsx}. Si es \code{NULL}, se crea un nuevo workbook.}

\item{nombre_hoja}{Nombre de la hoja. Si es \code{NULL}, se asigna un nombre basado en el valor de \code{hoja} (por ejemplo, "Hoja 1").}

\item{hoja}{Número o nombre de la hoja en la que se escribirá la información. Por defecto es 1.}

\item{formatear_tabla}{Lógico. Si es \code{TRUE}, se formatea la tabla utilizando \code{formatear_tabla_numerica} o \code{formatear_tabla_categorica} según el atributo \code{tipo_tabla} de \code{tabla}. Por defecto es \code{TRUE}.}

\item{tipo_tabla}{Carácter. Indica si la tabla es 'numerica' o 'categorica'. Si es \code{NULL}, se intenta detectar automáticamente a partir del atributo \code{tipo_tabla} de \code{tabla}.}

\item{hay_var_grupo}{Lógico. Indica si se ha utilizado una variable de agrupación. Si es \code{NULL}, se intenta detectar automáticamente a partir del atributo \code{vars_grupo} de \code{tabla}.}

\item{titulo_tabla}{Carácter. Título de la tabla. Si es \code{NULL}, se asigna el valor "Tabla 1".}

\item{numero_fila}{Número. La fila de inicio donde se escribirá la tabla. Por defecto es 4.}

\item{titulo_principal}{Carácter. Título principal que se escribirá en la parte superior de la hoja. Si es \code{NULL}, no se escribe título principal.}

\item{fuente}{Carácter. Nombre de la fuente a utilizar para los textos. Por defecto es "Calibri".}

\item{titulo_tabla_colorFuente}{Carácter. Color (en hexadecimal) para el texto del título de la tabla. Por defecto es "#00b2a9".}

\item{titulo_tabla_tamanyo}{Número. Tamaño de fuente para el título de la tabla. Por defecto es 12.}

\item{estilo_tabla}{Carácter. Estilo de tabla de Excel. Por defecto es "TableStyleLight1".}

\item{tabla_conFiltro}{Lógico. Si es \code{TRUE}, la tabla en Excel tendrá filtros. Por defecto es \code{TRUE}.}

\item{titulo_principal_tamanyo}{Número. Tamaño de fuente para el título principal. Por defecto es 14.}

\item{titulo_principal_colorFuente}{Carácter. Color (en hexadecimal) para el título principal. Por defecto es "#00b2a9".}

\item{nombre_archivo}{Carácter. Nombre del archivo Excel de salida (sin extensión). Si es \code{NULL}, se asigna "Tabla_analisisDescriptivo_temp".}

\item{exportar}{Lógico. Si es \code{TRUE}, se guarda el workbook en un archivo Excel. Por defecto es \code{TRUE}.}

\item{sobreescribir_archivo}{Lógico. Si es \code{TRUE}, se sobrescribe el archivo de salida en caso de existir. Por defecto es \code{TRUE}.}
}
\value{
Devuelve el objeto workbook con la hoja y la tabla agregada.
}
\description{
Esta función escribe una tabla (data.frame o tibble) en un objeto workbook de Excel, aplicando formato a los datos (por ejemplo, porcentajes y números) según el tipo de tabla (numérica o categórica). Además, permite agregar un título principal y un título para la tabla, y opcionalmente exporta el workbook a un archivo Excel.
}
\details{
La función valida que \code{tabla} sea un data.frame o tibble. Si se solicita formatear la tabla, se determina el tipo (numérica o categórica) a partir de los atributos \code{tipo_tabla} y \code{vars_grupo}. Se crea (o utiliza) un objeto workbook, se agrega una hoja y se escribe en ella el título principal (si se especifica), el título de la tabla y la tabla formateada. Si \code{exportar} es \code{TRUE}, el workbook se guarda en un archivo Excel.
}
\examples{
\dontrun{
  library(dplyr)
  temp <- mtcars \%>\%
   mutate(
   'cyl'=as.character(cyl),
   'carb'=factor(carb),
   'gear' = factor(gear),
   'grupo' = sample(c('A','B'),nrow(.),T),
   'w'=rlnorm(32))
 temp_num <- generar_descriptivo_numerico( datos = temp, var_numerica = 'mpg' )
 crear_Excel_tabla(temp_num, exportar = T)
}
}
