% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crear_hojaExcel.R
\name{crear_hojaExcel}
\alias{crear_hojaExcel}
\title{Genera una hoja en un workbook Excel con tablas y títulos formateados}
\usage{
crear_hojaExcel(
  list_tablas,
  titulo_principal = "DESCRIPTIVOS UNIVARIADOS",
  titulos_tablas,
  nombre_hoja = NULL,
  hoja = 1,
  estilo_tabla = "TableStyleLight1",
  nombre_archivo,
  exportar = FALSE,
  sobreescribir_archivo = TRUE,
  tabla_conFiltro = TRUE,
  workbook = NULL,
  fuente = "Calibri",
  titulo_principal_tamanyo = 14,
  titulo_principal_colorFuente = "#00b2a9",
  titulo_tabla_colorFuente = "#00b2a9",
  titulo_tabla_tamanyo = 12
)
}
\arguments{
\item{list_tablas}{Lista de data.frames que se exportarán.}

\item{titulo_principal}{Título principal que se escribirá en la fila 1.}

\item{titulos_tablas}{Vector con los títulos de cada tabla.}

\item{nombre_hoja}{(opcional) Nombre de la hoja en el workbook; si es \code{NULL} se genera a partir del número de hoja.}

\item{hoja}{Número de hoja en el workbook.}

\item{estilo_tabla}{Nombre del estilo de tabla predefinido de Excel (por ejemplo, "TableStyleLight1").}

\item{nombre_archivo}{Nombre base del archivo Excel a exportar (sin extensión).}

\item{exportar}{Lógico; si es \code{TRUE} guarda el archivo Excel.}

\item{sobreescribir_archivo}{Lógico; si es \code{TRUE} sobrescribe el archivo si ya existe.}

\item{tabla_conFiltro}{Lógico; si es \code{TRUE} la tabla se crea con filtros.}

\item{workbook}{(opcional) Objeto workbook de \code{openxlsx}; si es \code{NULL} se crea uno nuevo.}

\item{fuente}{Fuente a utilizar para los títulos.}

\item{titulo_principal_tamanyo}{Tamaño de fuente para el título principal.}

\item{titulo_principal_colorFuente}{Color de la fuente para el título principal.}

\item{titulo_tabla_colorFuente}{Color de la fuente para los títulos de las tablas.}

\item{titulo_tabla_tamanyo}{Tamaño de fuente para los títulos de las tablas.}
}
\value{
Un objeto \code{workbook} de \code{openxlsx} con la hoja y tablas agregadas.
}
\description{
Esta función escribe en una hoja de un workbook Excel un título principal, títulos para cada tabla
y las tablas propiamente dichas, aplicando estilos personalizados a los títulos.
}
\examples{
list_temp <- list(head(mtcars,5),head(mtcars,5),head(mtcars,5))

generar_hojaExcel(
 list_temp,
 titulo_principal = "DESCRIPTIVOS UNIVARIADOS",
 titulos_tablas = letters[1:length(list_temp)],
 nombre_archivo = 'prueba',
 exportar = TRUE
 )

}
