% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatear_tabla_numerica.R
\name{formatear_tabla_numerica}
\alias{formatear_tabla_numerica}
\title{Formatea una tabla numérica para Excel}
\usage{
formatear_tabla_numerica(
  tabla,
  precision_num = 0.1,
  escala_num = 1,
  sufijo_num = "",
  prefijo_num = "",
  estilo_positivo_num = "none",
  separador_decimal = ",",
  separador_miles = "."
)
}
\arguments{
\item{tabla}{Data.frame o tibble. La tabla que se desea formatear.}

\item{precision_num}{Número. Precisión para formatear los números. Por defecto es 0.1.}

\item{escala_num}{Número. Factor de escala para formatear los números. Por defecto es 1.}

\item{sufijo_num}{Carácter. Sufijo a añadir a los números. Por defecto es "".}

\item{prefijo_num}{Carácter. Prefijo a añadir a los números. Por defecto es "".}

\item{estilo_positivo_num}{Carácter. Estilo a aplicar a números positivos. Por defecto es "none".}

\item{separador_decimal}{Carácter. Separador decimal. Por defecto es ",".}

\item{separador_miles}{Carácter. Separador de miles. Por defecto es ".".}
}
\value{
Devuelve el data.frame o tibble con las columnas numéricas formateadas.
}
\description{
Esta función toma un data.frame o tibble y aplica formateo numérico a sus columnas.
Se formatean las columnas numéricas (exceptuando las columnas 'N', 'Missing' y 'p_value')
utilizando \code{scales::number}. Además, se formatea la columna \code{p_value} con mayor precisión.
}
\details{
Se utiliza \code{dplyr::mutate} junto con \code{dplyr::across} para aplicar la función
\code{scales::number} a las columnas numéricas, exceptuando aquellas identificadas como 'N', 'Missing' o 'p_value'.
La columna \code{p_value} se formatea utilizando una precisión mayor.
}
\examples{
\dontrun{
  df_formateado <- formatear_tabla_numerica(mi_dataframe)
}

}
