% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_descriptivo_categorico.R
\name{generar_descriptivo_categorico}
\alias{generar_descriptivo_categorico}
\title{Generar descriptivos para una variable categórica}
\usage{
generar_descriptivo_categorico(
  datos,
  var_categorica,
  vars_grupo = NULL,
  var_peso = NULL,
  pivot = TRUE,
  variable_pivot = c("var_grupo", "var_categorica"),
  nivel_confianza = 0.95,
  estrategia_valoresPerdidos = c("E", "A"),
  simplificar_output = TRUE
)
}
\arguments{
\item{datos}{Data frame o tibble que contiene la variable a analizar.}

\item{var_categorica}{Nombre (carácter) de la variable categórica.}

\item{vars_grupo}{Vector de nombres de variables de agrupación. Si \code{datos} ya está agrupado, se utilizarán
los grupos definidos en el objeto.}

\item{var_peso}{Nombre (carácter) de la variable de pesos. Si es \code{NULL} (por defecto), se realizan recuentos simples.}

\item{pivot}{Lógico. Si \code{TRUE} (por defecto) y el data frame está agrupado, se pivota la tabla (wide)
utilizando la variable de agrupación o la variable categórica, según lo definido en \code{variable_pivot}.}

\item{variable_pivot}{Vector de caracteres que indica la variable a utilizar para pivotear: \code{"var_grupo"} o
\code{"var_categorica"}. Por defecto es \code{c("var_grupo", "var_categorica")}.}

\item{nivel_confianza}{Nivel de confianza para los intervalos. Por defecto es \code{0.95}.}

\item{estrategia_valoresPerdidos}{Estrategia para el manejo de valores faltantes en la variable categórica. Se debe elegir
\code{"E"} para eliminar o \code{"A"} para agrupar (NS/NC). Por defecto es \code{c("E", "A")}, de modo que se selecciona
\code{"E"}.}

\item{simplificar_output}{Por defecto \code{TRUE}. Cuando se calculan descriptivos agrupados reduce la cantidad de información
exportada en caso de pivotar.}
}
\value{
Data frame con los descriptivos calculados para la variable categórica, que incluye:
\itemize{
\item \code{n}: Frecuencia (ponderada, si se especifica) de cada categoría.
\item \code{n_sinW}: Frecuencia sin ponderar (solo si se especifica la variable de peso).
\item \code{p}: Porcentaje calculado.
\item \code{p_Min} y \code{p_Max}: Límites inferior y superior del intervalo de confianza, ajustados con el nivel de confianza.
\item \code{sd}: Error estándar aproximado de la proporción.
}
}
\description{
Calcula las frecuencias, porcentajes y sus intervalos de confianza (Wilson) para una variable categórica,
opcionalmente agrupada y ponderada. Además, permite pivotear la tabla resultante a formato ancho si se está
trabajando con un tibble agrupado.
}
\details{
La función realiza los siguientes pasos:
\enumerate{
\item Se asegura que la variable categórica sea de tipo carácter, incluso si originalmente es numérica.
\item Si \code{datos} está agrupado, se utiliza la(s) variable(s) de agrupación.
\item Se cuentan las ocurrencias (ponderadas, si se especifica) de cada categoría.
\item Se calcula el porcentaje de cada categoría. En caso de tener valores faltantes, se aplica la estrategia seleccionada:
\describe{
\item{\code{"E"}}{Elimina las observaciones faltantes (no se contabilizan en los porcentajes).}
\item{\code{"A"}}{Agrupa los faltantes en la categoría \code{"NS/NC"}.}
}
\item Se ordena el resultado en función de la variable (descendente si es carácter, ascendente si es numérica).
\item Se calculan intervalos de confianza utilizando el método de Wilson (se requiere la función \code{ci_margenError_pWilson}).
\item Si \code{pivot} es \code{TRUE} y el data frame está agrupado, se pivota la tabla a formato ancho.
\item Se añaden sufijos en los nombres de los intervalos de confianza según el nivel de confianza.
}
}
\examples{
\dontrun{
  library(dplyr)
  # Sin agrupación y sin pesos
  mtcars$cyl <- factor(mtcars$cyl)
  mtcars$vs <- factor(mtcars$vs)
  resultado <- generar_descriptivo_categorico(
                 datos = mtcars,
                 var_categorica = "cyl",
                 nivel_confianza = 0.95,
                 estrategia_valoresPerdidos = "E"
               )

  # Con agrupación y pesos (suponiendo que 'w' es la variable de peso)
  mtcars$w <- rlnorm(32) # añadimos pesos
  resultado2 <- generar_descriptivo_categorico(
   datos = mtcars,
   var_categorica = "vs",
   vars_grupo = 'gear',
   var_peso = "w",
   pivot = TRUE,
   variable_pivot = "var_grupo",
   nivel_confianza = 0.95,
   estrategia_valoresPerdidos = "A"
  )
}

generar_descriptivo_categorico( temp, 'cyl' )
generar_descriptivo_categorico( temp, 'cyl', 'vs' )
generar_descriptivo_categorico( temp, 'cyl', var_peso = 'w' )
generar_descriptivo_categorico( temp, 'cyl', 'vs', var_peso = 'w' )

}
