% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_descriptivo_VDcategorica_varX.R
\name{generar_descriptivo_VDcategorica_varX}
\alias{generar_descriptivo_VDcategorica_varX}
\title{Genera un descriptivo para una variable dependiente categórica}
\usage{
generar_descriptivo_VDcategorica_varX(
  datos,
  var_VDcat,
  var_X,
  var_peso = NULL,
  variable_pivot = "var_grupo",
  estrategia_valoresPerdidos = "A",
  nivel_confianza = 0.95,
  simplificar_output = T
)
}
\arguments{
\item{datos}{Un data.frame o tibble con los datos.}

\item{var_VDcat}{El nombre de la variable dependiente categórica.}

\item{var_X}{El nombre de la variable X (puede ser numérica o categórica).}

\item{var_peso}{El nombre de la variable de pesos (opcional).}

\item{variable_pivot}{El nombre de la variable de agrupación para la pivoteo (por defecto "var_grupo").}

\item{estrategia_valoresPerdidos}{La estrategia para manejar los valores perdidos, por defecto 'A'.}

\item{nivel_confianza}{El nivel de confianza para los intervalos de confianza (por defecto, 0.95).}

\item{simplificar_output}{Indica si se debe simplificar el output (por defecto, TRUE).}
}
\value{
Un tibble con los resultados del análisis descriptivo, que incluye las métricas por grupos según el tipo de la variable X.
La tabla tendrá un atributo que indica el tipo de análisis ('categorica') y si la variable X es un grupo.
}
\description{
Esta función genera un resumen descriptivo para una variable dependiente categórica (\code{var_VDcat})
según los valores de una variable X, que puede ser numérica o categórica. En el caso de que la
variable X sea numérica, se utiliza la función \code{generar_descriptivo_numerico} para generar el descriptivo
por grupos. Si la variable X es categórica, se utiliza la función \code{generar_descriptivo_categorico}.
}
\details{
Si se proporciona una variable de peso (\code{var_peso}), las métricas se ajustan en función de los pesos.
}
\examples{
\dontrun{
 library(analisisDescriptivo)
 library(dplyr)
 temp <- mtcars \%>\%
   mutate('cyl'=as.character(cyl),
          'carb'=factor(carb),
          'gear' = factor(gear),
          'grupo' = sample(c('A','B'),nrow(.),T),
          'w'=rlnorm(32),
          w_norm = rnorm(32)
   )

 generar_descriptivo_VDcategorica_varX(
   temp,
   'gear',
   'mpg',
   'w'
 )
 generar_descriptivo_VDcategorica_varX(
   temp,
   'gear',
   'grupo',
   'w'
 )
}

export
}
