% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_descriptivo_VDnumerica_varX.R
\name{generar_descriptivo_VDnumerica_varX}
\alias{generar_descriptivo_VDnumerica_varX}
\title{Genera un descriptivo para una variable dependiente numérica por grupos definidos por una variable X}
\usage{
generar_descriptivo_VDnumerica_varX(
  datos,
  var_VDnum,
  var_X,
  var_peso = NULL,
  nivel_confianza = 0.95
)
}
\arguments{
\item{datos}{Un data.frame o tibble con los datos.}

\item{var_VDnum}{El nombre de la variable dependiente numérica.}

\item{var_X}{El nombre de la variable X (puede ser numérica o categórica).}

\item{var_peso}{El nombre de la variable de pesos (opcional).}

\item{nivel_confianza}{El nivel de confianza para los intervalos de confianza (por defecto, 0.95).}
}
\value{
Un tibble con los resultados del análisis descriptivo, que incluye la media por cuartiles (si X es numérica),
el coeficiente de la regresión OLS, el valor p de la regresión, y otros detalles de la regresión ponderada si se proporciona \code{var_peso}.
}
\description{
Esta función genera un resumen descriptivo para una variable dependiente numérica (\code{var_VDnum})
según los valores de una variable X, que puede ser numérica o categórica. En el caso de que la
variable X sea numérica, se calcula la media por cuartiles y se ajusta un modelo de regresión lineal
(OLS). Si la variable X es categórica, se utiliza la función \code{generar_descriptivo_numerico} para
generar el descriptivo por grupos.
}
\details{
Si se proporciona una variable de peso (\code{var_peso}), se calcula también el modelo de regresión ponderada
y se ajustan las métricas correspondientes.
}
\examples{
\dontrun{
 library(analisisDescriptivo)
 library(dplyr)
 temp <- mtcars \%>\%
   mutate('cyl'=as.character(cyl),
          'carb'=factor(carb),
          'gear' = factor(gear),
          'grupo' = sample(c('A','B'),nrow(.),T),
          'w'=rlnorm(32),
          w_norm = rnorm(32)
   )

 generar_descriptivo_VDnumerica_varX(
   temp,
   'mpg',
   'disp',
   'w'
 )
 generar_descriptivo_VDnumerica_varX(
   temp,
   'mpg',
   'grupo',
   'w'
 )
}

}
