% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_descriptivos_univariados.R
\name{generar_descriptivos_univariados}
\alias{generar_descriptivos_univariados}
\title{Generar descriptivos univariados}
\usage{
generar_descriptivos_univariados(
  datos,
  vars_categoricas = NULL,
  vars_numericas = NULL,
  var_peso = NULL,
  nivel_confianza = 0.95,
  selecc_vars_auto = TRUE,
  pivot = TRUE,
  variable_pivot = c("var_grupo", "var_categorica"),
  estrategia_valoresPerdidos = c("E", "A"),
  digits = 0.1,
  num_unificar_1tabla = FALSE
)
}
\arguments{
\item{datos}{Data frame o tibble con los datos a analizar.}

\item{vars_categoricas}{Vector de nombres de variables categóricas. Si es \code{NULL} y \code{selecc_vars_auto} es \code{TRUE},
se seleccionan automáticamente las variables de tipo carácter o factor.}

\item{vars_numericas}{Vector de nombres de variables numéricas. Si es \code{NULL} y \code{selecc_vars_auto} es \code{TRUE},
se seleccionan automáticamente las variables numéricas.}

\item{var_peso}{Nombre de la variable de peso, si se desea ponderar los cálculos. Por defecto \code{NULL}.}

\item{nivel_confianza}{Nivel de confianza para los intervalos. Por defecto \code{0.95}.}

\item{selecc_vars_auto}{Lógico. Si es \code{TRUE}, se seleccionan automáticamente las variables si no se especifican. Por defecto \code{TRUE}.}

\item{pivot}{Lógico. Si es \code{TRUE}, se pivota la salida de variables categóricas a formato ancho (solo si los datos están agrupados). Por defecto \code{TRUE}.}

\item{variable_pivot}{Vector de caracteres que indica la variable a utilizar para pivotear: \code{"var_grupo"} o \code{"var_categorica"}. Por defecto \code{"var_grupo"}.}

\item{estrategia_valoresPerdidos}{Estrategia para el manejo de valores faltantes en variables categóricas: \code{"E"} (eliminar) o \code{"A"} (agrupar en "NS/NC"). Por defecto \code{"E")}.}

\item{digits}{Valor de precisión para formatear los descriptivos numéricos. Por defecto \code{0.1}.}

\item{num_unificar_1tabla}{Lógico. Si es \code{TRUE}, se retorna un único data frame combinando los descriptivos numéricos; si es \code{FALSE},
se retorna una lista. Por defecto \code{FALSE}. Sólo para las numéricas}
}
\value{
Una lista con dos elementos:
\itemize{
\item \code{Numericas}: Descriptivos para variables numéricas.
\item \code{Categoricas}: Descriptivos para variables categóricas.
}
}
\description{
Calcula descriptivos univariados para variables categóricas y numéricas de un data frame o tibble.
Si no se especifican las variables, se seleccionan automáticamente según su clase.
}
\details{
La función aplica internamente \code{generar_descriptivos_categoricos} y \code{generar_descriptivos_numericos}.
}
\examples{
\dontrun{
  # Ejemplo con variables especificadas
  resultados <- generar_descriptivos_univariados(
                  datos = mtcars,
                  vars_categoricas = c("cyl", "gear"),
                  vars_numericas = c("mpg", "hp"),
                  nivel_confianza = 0.95,
                  pivot = TRUE,
                  variable_pivot = "var_grupo",
                  estrategia_valoresPerdidos = "E",
                  digits = 0.1,
                  num_unificar_1tabla = FALSE
                )
}

}
