% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generar_descriptivos_VDnumerica.R
\name{generar_descriptivos_VDnumerica}
\alias{generar_descriptivos_VDnumerica}
\title{Genera descriptivos para una variable dependiente numérica por múltiples variables X}
\usage{
generar_descriptivos_VDnumerica(
  datos,
  var_VDnum,
  vars_X = NULL,
  var_peso = NULL,
  num_unificar_1tabla = T,
  nivel_confianza = 0.95,
  selecc_vars_auto = FALSE,
  estrategia_valoresPerdidos_grupo = c("Nada", "Agrupar", "Eliminar")
)
}
\arguments{
\item{datos}{Un data.frame o tibble con los datos.}

\item{var_VDnum}{El nombre de la variable dependiente numérica.}

\item{vars_X}{Una lista de nombres de variables X, que pueden ser numéricas o categóricas. Si es NULL, se seleccionan automáticamente las demás columnas.}

\item{var_peso}{El nombre de la variable de pesos (opcional).}

\item{num_unificar_1tabla}{Un valor lógico que indica si los resultados numéricos deben unificarse en una sola tabla (por defecto TRUE).}

\item{nivel_confianza}{El nivel de confianza para los intervalos de confianza (por defecto, 0.95).}

\item{selecc_vars_auto}{Un valor lógico que indica si se deben seleccionar automáticamente las variables X si no se especifican (por defecto FALSE).}

\item{estrategia_valoresPerdidos_grupo}{Estrategia para el manejo de valores faltantes en la variable de agrupación. Por defecto \code{"Nada"}. \code{"Eliminar"} para eliminar o \code{"Agrupar"} para agrupar (NS/NC).}
}
\value{
Un tibble con los resultados del análisis descriptivo, que incluye tanto los descriptivos para las variables X numéricas como categóricas.
}
\description{
Esta función genera un resumen descriptivo para una variable dependiente numérica (\code{var_VDnum})
con respecto a varias variables X. Las variables X pueden ser tanto numéricas como categóricas.
Los resultados se calculan utilizando la función \code{generar_descriptivo_VDnumerica} para cada variable X.
Además, si se proporciona una variable de peso (\code{var_peso}), los análisis se realizarán ponderados.
}
\details{
Si no se proporciona la lista de variables X, la función automáticamente seleccionará todas las variables
en \code{datos} que no sean \code{var_VDnum} ni \code{var_peso} y las tratará como posibles variables X.
}
\examples{
library(analisisDescriptivo)
temp <- mtcars \%>\%
  mutate('cyl'=as.character(cyl),
         'carb'=factor(carb),
         'gear' = factor(gear),
         'grupo' = sample(c('A','B'),nrow(.),T),
         'w'=rlnorm(32),
         w_norm = rnorm(32)
  )
 generar_descriptivos_VDnumerica(
   datos = temp,
   var_VDnum = 'mpg',
   vars_X = c('grupo')
 )
}
