% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalo_confianza_pWilson.R
\name{intervalo_confianza_pWilson}
\alias{intervalo_confianza_pWilson}
\title{Calcular límite del intervalo de confianza para proporciones usando el método de Wilson}
\usage{
intervalo_confianza_pWilson(
  p,
  N,
  nivel_confianza = 0.95,
  limite = c("inferior", "superior")
)
}
\arguments{
\item{p}{Vector de proporciones (valores entre 0 y 1).}

\item{N}{Vector con el tamaño de muestra correspondiente a cada proporción.}

\item{nivel_confianza}{Nivel de confianza. Por defecto es \code{0.95}.}

\item{limite}{Cadena que indica si se desea el límite inferior (\code{"inferior"}) o superior (\code{"superior"}) del intervalo.}
}
\value{
Vector numérico con el límite inferior o superior del intervalo de confianza para cada par de \code{p} y \code{N}.
}
\description{
Calcula el límite inferior o superior del intervalo de confianza para proporciones utilizando el método de Wilson.
}
\details{
Se utiliza la función \code{qnorm} para obtener el valor crítico de la distribución normal y
se calcula el intervalo de confianza siguiendo la fórmula de Wilson.
}
\examples{
intervalo_confianza_pWilson(c(0.1, 0.5), c(100, 200), nivel_confianza = 0.95, limite = "superior")

}
