% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatear_numero.R
\name{number}
\alias{number}
\title{Formatear números con separadores}
\usage{
number(x, accuracy = 0.1, ...)
}
\arguments{
\item{x}{Vector numérico.}

\item{accuracy}{Precisión para el formateo del número. Por defecto es \code{0.1}.}

\item{...}{Argumentos adicionales que se pasan a \code{scales::number}.}
}
\value{
Vector de caracteres con los números formateados.
}
\description{
Convierte un vector numérico a una representación en formato numérico legible, utilizando la función \code{scales::number}.
}
\examples{
number(c(1234.56, 7890.12))

}
