% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_diferencia_medias.R
\name{obtener_diferencia_medias}
\alias{obtener_diferencia_medias}
\title{Función interna: extraer las diferencias de medias}
\usage{
obtener_diferencia_medias(model_lm, var_grupo)
}
\arguments{
\item{model_lm}{Modelo lineal calculado con \code{lm}}

\item{var_grupo}{Variable categórica incluida como variabe independiente.}
}
\value{
Un tible con los coeficientes y el p.value.
}
\description{
Función de uso interno en la función \(code){generar_descriptivo_numerico}. Tras haber
calculado el modelo lineal OLS, esta función se utiliza para extraer si exiten diferencias
significativas entre las medias de los grupos.
}
\details{
Sólo se utiliza cuando se agrupa por una variable categórica.
}
