% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_formula.R
\name{obtener_formula}
\alias{obtener_formula}
\title{Obtener Fórmula para Modelos Lineales}
\usage{
obtener_formula(VD, Xs)
}
\arguments{
\item{VD}{Character. Nombre de la variable dependiente.}

\item{Xs}{Character vector. Nombre(s) de la(s) variable(s) independiente(s).}
}
\value{
Un objeto de clase \code{formula} en el formato \code{VD ~ X1 + X2 + ...}.
}
\description{
Genera una fórmula en formato R a partir de una variable dependiente y un vector de variables independientes.
}
\details{
La función utiliza \code{paste0} para concatenar los nombres de las variables independientes,
generando así una fórmula que puede usarse en funciones como \code{lm()}.
}
\examples{
# Generar una fórmula para ajustar un modelo lineal:
formula_modelo <- obtener_formula("mpg", c("cyl", "hp"))
print(formula_modelo)

}
