% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatear_porcentaje.R
\name{percent}
\alias{percent}
\title{Convertir valores numéricos a formato porcentual}
\usage{
percent(x, accuracy = 1, ...)
}
\arguments{
\item{x}{Vector numérico.}

\item{accuracy}{Precisión para el formateo del porcentaje. Por defecto es \code{1}.}

\item{...}{Argumentos adicionales que se pasan a \code{scales::percent}.}
}
\value{
Vector de caracteres con los porcentajes formateados.
}
\description{
Convierte un vector numérico a formato de porcentaje utilizando la función \code{scales::percent}.
}
\examples{
percent(c(0.1, 0.5, 0.99), 0.01)

}
