% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preparar_tabla_excel.R
\name{preparar_tabla_excel}
\alias{preparar_tabla_excel}
\title{Formatear tabla para exportación a Excel}
\usage{
preparar_tabla_excel(datos, ajustar_proporciones = NULL, precision = 0.1)
}
\arguments{
\item{datos}{Data frame o tibble que se desea formatear.}

\item{ajustar_proporciones}{Vector de caracteres con el nombre de las columnas que representan proporciones.
Si es \code{NULL} (por defecto), no se aplica ajuste.}

\item{precision}{Precisión para el formateo de las proporciones. Por defecto es \code{0.1}.}
}
\value{
Data frame formateado, en el cual la primera fila contiene los nombres originales de las columnas,
y las columnas han sido convertidas a caracteres.
}
\description{
Prepara una tabla (tibble o data frame) para exportarla a una única hoja de Excel.
La función formatea las columnas indicadas como proporciones (si se especifica) usando la función
\code{percent} y convierte todas las columnas a caracteres. Además, conserva los nombres originales
de las columnas en la primera fila, mientras renombra internamente las columnas a un formato genérico.
}
\details{
La función realiza los siguientes pasos:
\enumerate{
\item Si se especifica \code{ajustar_proporciones}, se aplica la función \code{percent} a esas columnas.
\item Se convierten todas las columnas a caracteres.
\item Se extraen los nombres originales de las columnas y se guardan en un vector.
\item Se renombran las columnas a "col1", "col2", ..., para luego insertar la fila de encabezado.
\item Se une la fila con los nombres originales y la tabla formateada.
}
}
\note{
\strong{Deprecated} Esta función está obsoleta y será eliminada en futuras actualizaciones. Utiliza las funciones \verb{crear_Excel_*()} en su lugar.
}
\examples{
\dontrun{
  library(dplyr)
  library(tibble)
  datos <- tibble(a = c(0.1, 0.2, 0.3), b = c(1, 2, 3))
  tabla_formateada <- preparar_tabla_excel(datos, ajustar_proporciones = "a", precision = 1)
}

}
