% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodos_s3.R
\name{print.tabla_descript}
\alias{print.tabla_descript}
\title{Impresion de los resultados de los analisis}
\usage{
\method{print}{tabla_descript}(x, ...)
}
\arguments{
\item{x}{Objeto de la clase tabla_descript}

\item{...}{Argumentos adicionales}
}
\value{
Devuelve de forma invisible el objeto x
}
\description{
Metodo print para mostrar los resultados obtenidos de la funcion principal
}
