% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desviacion_estandar_ponderada.R
\name{sd_pond}
\alias{sd_pond}
\title{Calcular la desviación estándar ponderada}
\usage{
sd_pond(x, pesos)
}
\arguments{
\item{x}{Vector numérico.}

\item{pesos}{Vector de pesos del mismo tamaño que \code{x}.}
}
\value{
La desviación estándar ponderada.
}
\description{
Calcula la desviación estándar de un vector numérico \code{x} utilizando un vector de pesos \code{w}.
}
\details{
Se eliminan los valores \code{NA} de \code{x} y sus correspondientes pesos en \code{pesos}. La función utiliza
la corrección de Bessel (dividiendo por \code{sum(pesos)-1}) para el cálculo de la varianza.
}
\examples{
sd_pond(c(1,2,3,4,5), c(1,1,1,1,1))

}
