% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{suma_valores}
\alias{suma_valores}
\title{Suma de todos los valores de las columnas numericas del data frame}
\usage{
suma_valores(tabla, columnas = NULL)
}
\arguments{
\item{tabla}{Data frame con varias columnas numericas para cada muestra}

\item{columnas}{Valor opcional \code{columnas} para filtrar por la(s) variable(s)
que se quieren analizar}
}
\value{
Data frame con el \code{n_validos} y \code{n_NAs} correspondiente a cada columna
}
\description{
Calcula la suma de la cantidad de valores validos como perdidos
}
