% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcion_principal.R
\name{tabla_descript}
\alias{tabla_descript}
\title{Tabla con los resultados descriptivos}
\usage{
tabla_descript(tabla, columnas = NULL, nombre_col = NULL, val_col = NULL)
}
\arguments{
\item{tabla}{Data  frame a analizar con columnas numericas para las muestras
y columnas adicionales (numericas o caracteres) para filtrar}

\item{columnas}{Valor opcional que corresponde con una lista con el nombre de
las columnas que se quieren analizar. Si no se proporciona un valor, se analizan
todas las columnas numericas}

\item{nombre_col}{Valor opcional para filtrar los datos en base a la columna
deseada}

\item{val_col}{Valor opcional para filtrar por el valor deseado de la columna
seleccionada previamente}
}
\value{
Objeto de clase \code{tabla_descript} que contiene la tabla con los
estadisticos descriptivos, la tabla de frecuencia, los parametros del analisis
y las advertencias (si fuera el caso)
}
\description{
Agrupa los resultados de las otras funciones para mostrar todos los resultados
descriptivos con opcion a filtrarlos segun las necesidades del usuario
}
