% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metodos_s3.R
\name{tabla_descriptiva}
\alias{tabla_descriptiva}
\title{Objeto S3 de resultados del analisis}
\usage{
tabla_descriptiva(resultados, parametros, advertencias = character())
}
\arguments{
\item{resultados}{Lista con los resultados de los analisis de la funcion
principal}

\item{parametros}{Lista con los parametros usados durante el analisis}

\item{advertencias}{Vector de caracteres con las advertencias producidas
durante el analisis (si las hubiera)}
}
\value{
Objeto de la clase tabla_descript
}
\description{
Constructor S3 que muestra los resultados de la funcion principal
}
