% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trocear_por_cuartiles.R
\name{trocear_por_cuartiles}
\alias{trocear_por_cuartiles}
\title{Trocea un vector numérico en cuartiles}
\usage{
trocear_por_cuartiles(x, añadirValor = F, ...)
}
\arguments{
\item{x}{Un vector numérico.}

\item{añadirValor}{Por defecto FALSE. Si TRUE añade a la etiqueta el valor del cuartil.}

\item{...}{Parámetros de la función \code{analisisDescriptivo::formatear_numero}.}
}
\value{
Un vector con las etiquetas de cuartiles ('Cuartil_1', 'Cuartil_2', 'Cuartil_3', 'Cuartil_4')
}
\description{
Esta función toma un vector numérico y lo divide en cuatro cuartiles. Cada elemento del vector
es asignado a un cuartil dependiendo de su valor.
}
\examples{
trocear_por_cuartiles(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
trocear_por_cuartiles(mtcars$mpg)
trocear_por_cuartiles(mtcars$mpg, añadirValor = T)
}
