% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trocear_por_cuartiles.R
\name{trocear_por_cuartiles}
\alias{trocear_por_cuartiles}
\title{Trocea un vector numérico en cuartiles}
\usage{
trocear_por_cuartiles(x)
}
\arguments{
\item{x}{Un vector numérico.}
}
\value{
Un vector con las etiquetas de cuartiles ('Cuartil_1', 'Cuartil_2', 'Cuartil_3', 'Cuartil_4')
}
\description{
Esta función toma un vector numérico y lo divide en cuatro cuartiles. Cada elemento del vector
es asignado a un cuartil dependiendo de su valor.
}
\examples{
trocear_por_cuartiles(c(1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}
